requirements.txt
requests
rapidfuzz
pandas
numpy
jinja2
python-dotenv
